Abstract selectedKeys and method to utils
